test: satisfy root composition lint gates
